requirements.txt

openai~=1.74.0
python-dotenv~=1.1.0